push current wip progress

+++ include/NeuralNetwork.h
@@ -12,7 +12,7 @@ class NeuralNetwork {
 public:
   NeuralNetwork(std::unique_ptr<LossFunction> loss_function, int iterations, int divisor);
   void add_layer(std::unique_ptr<BaseLayer> layer);
-  void train_network(const Eigen::MatrixXf &input, const Eigen::MatrixXf &label);
+  void train_network(const Eigen::MatrixXf &input, const Eigen::MatrixXi &label);
   const Eigen::MatrixXf &test_network(const Eigen::MatrixXf &input);
   Eigen::MatrixXf get_result();
 
@@ -22,7 +22,7 @@ private:
 
 
   void feed_forward(const Eigen::MatrixXf &input);
-  void backpropagation(const Eigen::MatrixXf &input, const Eigen::MatrixXf &label);
+  void backpropagation(const Eigen::MatrixXf &input, const Eigen::MatrixXi &label);
   void update();
   int m_iterations;
   int m_divisor;
